Introduce `TokenRefreshFlow` type
